# src/main/resources/openapi.yaml
openapi: 3.1.0
info:
  title: Wallet API
  description: Personal finance management REST API for tracking accounts, transactions, budgets, and goals
  version: 1.0.0
  contact:
    name: Wallet
    url: https://github.com/Americo91/wallet
servers:
  - url: http://localhost:8080
    description: Local development server
...
paths:
...
  /api/v1/categories:
    get:
      tags: [Categories]
      operationId: listCategories
      summary: List categories
      parameters:
        - name: name
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CategoryDto"
    post:
      tags: [Categories]
      operationId: createCategory
      summary: Create a new category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryDto"
      responses:
        "201":
          description: Category created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryDto"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
  /api/v1/categories/{categoryId}:
    get:
      tags: [Categories]
      operationId: getCategoryById
      summary: Get category by ID
      parameters:
        - $ref: "#/components/parameters/CategoryId"
      responses:
        "200":
          description: Category found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryDto"
        "400":
          description: Category not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
    put:
      tags: [Categories]
      operationId: updateCategory
      summary: Update a category
      parameters:
        - $ref: "#/components/parameters/CategoryId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryDto"
      responses:
        "204":
          description: Category updated
        "400":
          description: Validation error or category not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
    delete:
      tags: [Categories]
      operationId: deleteCategory
      summary: Delete a category
      parameters:
        - $ref: "#/components/parameters/CategoryId"
      responses:
        "204":
          description: Category deleted
        "400":
          description: Category not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"

  # ── Labels ────────────────────────────────────────────────────
...
  /api/v1/budgets:
    get:
      tags: [Budgets]
      operationId: listBudgets
      summary: List all budgets
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BudgetDto"
    post:
      tags: [Budgets]
      operationId: createBudget
      summary: Create a new budget
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BudgetDto"
      responses:
        "201":
          description: Budget created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BudgetDto"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
  /api/v1/budgets/{budgetId}:
    get:
      tags: [Budgets]
      operationId: getBudgetById
      summary: Get budget by ID
      parameters:
        - $ref: "#/components/parameters/BudgetId"
      responses:
        "200":
          description: Budget found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BudgetDto"
        "400":
          description: Budget not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
    put:
      tags: [Budgets]
      operationId: updateBudget
      summary: Update a budget
      parameters:
        - $ref: "#/components/parameters/BudgetId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BudgetDto"
      responses:
        "204":
          description: Budget updated
        "400":
          description: Validation error or budget not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
    delete:
      tags: [Budgets]
      operationId: deleteBudget
      summary: Delete a budget
      parameters:
        - $ref: "#/components/parameters/BudgetId"
      responses:
        "204":
          description: Budget deleted
        "400":
          description: Budget not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"

  # ── Goals ─────────────────────────────────────────────────────
  /api/v1/goals:
    get:
      tags: [Goals]
      operationId: listGoals
      summary: List goals
      parameters:
        - name: name
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GoalDto"
    post:
      tags: [Goals]
      operationId: createGoal
      summary: Create a new goal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoalDto"
      responses:
        "201":
          description: Goal created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GoalDto"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
  /api/v1/goals/{goalId}:
    get:
      tags: [Goals]
      operationId: getGoalById
      summary: Get goal by ID
      parameters:
        - $ref: "#/components/parameters/GoalId"
      responses:
        "200":
          description: Goal found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GoalDto"
        "400":
          description: Goal not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
    put:
      tags: [Goals]
      operationId: updateGoal
      summary: Update a goal
      parameters:
        - $ref: "#/components/parameters/GoalId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoalDto"
      responses:
        "204":
          description: Goal updated
        "400":
          description: Validation error or goal not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
    delete:
      tags: [Goals]
      operationId: deleteGoal
      summary: Delete a goal
      parameters:
        - $ref: "#/components/parameters/GoalId"
      responses:
        "204":
          description: Goal deleted
        "400":
          description: Goal not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
  /api/v1/goals/{goalId}/addSavedAmount:
    post:
      tags: [Goals]
      operationId: addSavedAmount
      summary: Add to a goal's saved amount
      parameters:
        - $ref: "#/components/parameters/GoalId"
      requestBody:
        required: true
        description: Amount to add to saved total
        content:
          application/json:
            schema:
              type: number
              format: double
              multipleOf: 0.01
      responses:
        "200":
          description: Updated goal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GoalDto"
        "400":
          description: Goal not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
  /api/v1/goals/{goalId}/markAsSolved:
    put:
      tags: [Goals]
      operationId: markGoalAsReached
      summary: Mark a goal as reached
      parameters:
        - $ref: "#/components/parameters/GoalId"
      responses:
        "204":
          description: Goal marked as reached
        "400":
          description: Goal not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"

  # ── Standing Orders ───────────────────────────────────────────
...
  /api/v1/standing-orders/upcoming:
    get:
      tags: [Standing Orders]
      operationId: getUpcomingStandingOrders
      summary: Get upcoming standing orders
      parameters:
        - name: days
          in: query
          required: false
          description: Number of days to look ahead
          schema:
            type: integer
            format: int32
            default: 3
      responses:
        "200":
          description: Upcoming standing orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StandingOrderDto"
  /api/v1/standing-orders/{standingOrderId}:
    get:
      tags: [Standing Orders]
      operationId: getStandingOrderById
      summary: Get standing order by ID
      parameters:
        - $ref: "#/components/parameters/StandingOrderId"
      responses:
        "200":
          description: Standing order found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StandingOrderDto"
        "400":
          description: Standing order not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
    put:
      tags: [Standing Orders]
      operationId: updateStandingOrder
      summary: Update a standing order
      parameters:
        - $ref: "#/components/parameters/StandingOrderId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StandingOrderDto"
      responses:
        "204":
          description: Standing order updated
        "400":
          description: Validation error or standing order not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
    delete:
      tags: [Standing Orders]
      operationId: deleteStandingOrder
      summary: Delete a standing order
      parameters:
        - $ref: "#/components/parameters/StandingOrderId"
      responses:
        "204":
          description: Standing order deleted
        "400":
          description: Standing order not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
  /api/v1/accounts/{accountId}/standing-orders:
    post:
      tags: [Standing Orders]
      operationId: createStandingOrder
      summary: Create a new standing order for an account
      parameters:
        - $ref: "#/components/parameters/AccountId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StandingOrderDto"
      responses:
        "201":
          description: Standing order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StandingOrderDto"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
components:
  parameters:
...
    AccountId:
      name: accountId
      in: path
      required: true
      description: Account UUID
      schema:
        type: string
        format: uuid
...
    CategoryId:
      name: categoryId
      in: path
      required: true
      description: Category UUID
      schema:
        type: string
        format: uuid
...
    BudgetId:
      name: budgetId
      in: path
      required: true
      description: Budget UUID
      schema:
        type: string
        format: uuid
    GoalId:
      name: goalId
      in: path
      required: true
      description: Goal UUID
      schema:
        type: string
        format: uuid
    StandingOrderId:
      name: standingOrderId
      in: path
      required: true
      description: Standing order UUID
      schema:
        type: string
        format: uuid
  schemas:
    ErrorDto:
      type: object
      properties:
        message:
          type: string
          example: "Institution not found"
      required:
        - message
...
    CategoryDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
          example: "d4e5f6a7-b8c9-0123-defa-234567890123"
        name:
          type: string
          example: "Groceries"
        type:
          type: string
          enum: [EXPENSE, INCOME, TRANSFER]
          default: EXPENSE
          example: "EXPENSE"
        parentId:
          type: [string, "null"]
          format: uuid
          example: "f6a7b8c9-d0e1-2345-abcd-456789012345"
        createdAt:
          type: string
          format: date-time
          readOnly: true
          example: "2025-01-10T08:00:00Z"
        updatedAt:
          type: string
          format: date-time
          readOnly: true
          example: "2025-01-10T08:00:00Z"
      required:
        - name
...
    BudgetDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
          example: "f6a7b8c9-d0e1-2345-abcd-456789012345"
        name:
          type: string
          example: "Monthly Groceries"
        period:
          type: string
          enum: [DAILY, WEEKLY, MONTHLY, YEARLY, UNA_TANTUM]
          example: "MONTHLY"
        budgetLimit:
          type: number
          example: 400.00
        categoryIds:
          type: [array, "null"]
          items:
            type: string
            format: uuid
          uniqueItems: true
          example: ["d4e5f6a7-b8c9-0123-defa-234567890123"]
        accountIds:
          type: [array, "null"]
          items:
            type: string
            format: uuid
          uniqueItems: true
          example: ["b2c3d4e5-f6a7-8901-bcde-f12345678901"]
        labelIds:
          type: [array, "null"]
          items:
            type: string
            format: uuid
          uniqueItems: true
          example: ["e5f6a7b8-c9d0-1234-efab-345678901234"]
        closed:
          type: boolean
          example: false
        overLimit:
          type: [array, "null"]
          items:
            $ref: "#/components/schemas/Overlimit"
          uniqueItems: true
          readOnly: true
        createdAt:
          type: string
          format: date-time
          readOnly: true
          example: "2025-02-01T00:00:00Z"
        updatedAt:
          type: string
          format: date-time
          readOnly: true
          example: "2025-02-01T00:00:00Z"
      required:
        - name
        - period
        - budgetLimit
        - categoryIds
        - accountIds
        - labelIds
    Overlimit:
      type: object
      properties:
        limit:
          type: number
          example: 400.00
        periodAmount:
          type: number
          example: 450.30
        periodStart:
          type: string
          format: date
          example: "2025-06-01"
        periodEnd:
          type: string
          format: date
          example: "2025-06-30"
    GoalDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
          example: "a7b8c9d0-e1f2-3456-bcde-567890123456"
        name:
          type: string
          example: "Emergency Fund"
        amount:
          type: number
          default: 0
          example: 10000.00
        savedAmount:
          type: number
          default: 0
          example: 3500.00
        reached:
          type: boolean
          example: false
        createdAt:
          type: string
          format: date-time
          readOnly: true
          example: "2025-01-01T00:00:00Z"
        updatedAt:
          type: string
          format: date-time
          readOnly: true
          example: "2025-06-15T12:00:00Z"
      required:
        - name
        - amount
    StandingOrderDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
          example: "b8c9d0e1-f2a3-4567-cdef-678901234567"
        name:
          type: string
          example: "Spotify"
        amount:
          type: number
          example: 9.99
        currency:
          type: string
          enum: [EUR, USD, JPY, ALL]
          example: "EUR"
        frequency:
          type: string
          enum: [DAILY, WEEKLY, MONTHLY, YEARLY, UNA_TANTUM]
          example: "MONTHLY"
        type:
          type: string
          enum: [EXPENSE, INCOME]
          example: "EXPENSE"
        nextOccurrence:
          type: string
          format: date
          example: "2025-07-01"
        account:
          type: string
          format: uuid
          readOnly: true
          example: "b2c3d4e5-f6a7-8901-bcde-f12345678901"
        category:
          type: string
          format: uuid
          example: "d4e5f6a7-b8c9-0123-defa-234567890123"
        labels:
          type: [array, "null"]
          items:
            type: string
            format: uuid
          uniqueItems: true
          example: ["e5f6a7b8-c9d0-1234-efab-345678901234"]
        description:
          type: [string, "null"]
          example: "Spotify Premium subscription"
        payee:
          type: [string, "null"]
          example: "Spotify AB"
        enabled:
          type: boolean
          example: true
        createdAt:
          type: string
          format: date-time
          readOnly: true
          example: "2025-03-01T09:00:00Z"
        updatedAt:
          type: string
          format: date-time
          readOnly: true
          example: "2025-03-01T09:00:00Z"
      required:
        - name
        - amount
        - currency
        - frequency
        - type
        - nextOccurrence
        - category